backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-dotenv
cryptography
passlib[bcrypt]
PyJWT
python-decouple
python-multipart==0.0.16
email-validator==2.2.0